+++ src/shared/webview-commands.ts
@@ -183,6 +183,7 @@ export class AddOntologyItemCommand {
 export class CreateNodeCommand {
 	public readonly type = 'createNode';
 	public readonly payload?: ModelTreeItemDropPayload;
+	public readonly payloads?: readonly ModelTreeItemDropPayload[];
 	public readonly position: CanvasPoint;
 	public readonly size?: {
 		readonly width: number;
@@ -191,6 +192,7 @@ export class CreateNodeCommand {
 
 	public constructor(options: {
 		readonly payload?: ModelTreeItemDropPayload;
+		readonly payloads?: readonly ModelTreeItemDropPayload[];
 		readonly position: CanvasPoint;
 		readonly size?: {
 			readonly width: number;
@@ -198,6 +200,7 @@ export class CreateNodeCommand {
 		};
 	}) {
 		this.payload = options.payload;
+		this.payloads = options.payloads;
 		this.position = options.position;
 		this.size = options.size;
 	}